Removed class definitions, imported classes, and renamed animals.py to index.py
